Extract create collection method to helper
