+++ spec.features/collection_spec.rb
@@ -23,18 +23,6 @@ describe 'Collections:' do
   let(:file_1) { GenericFile.first }
   let(:file_2) { GenericFile.last }
 
-  def create_collection (title, creator, description)
-    page.should have_content 'Create New Collection'
-    fill_in('Title', with: title)
-    fill_in('Creator', with: creator)
-    fill_in('Description', with: description)
-    click_button("Create Collection")
-    page.should have_content 'Items in this Collection'
-    page.should have_content(title)
-    page.should have_content(creator)
-    page.should have_content(description)
-  end
-
   describe 'When creating a collection:' do
 
     specify "I should be able to create it without any files" do
