+++ spec.features/feature_spec_helper.rb
@@ -9,4 +9,5 @@ require_relative './support/helpers/fixtures'
 require_relative './support/mail'
 require_relative './support/helpers/generic_files'
 require_relative './support/selectors'
+require_relative './support/helpers/collections'
