+++ spec.features/support/helpers/collections.rb
@@ -1,0 +1,17 @@
+module CollectionsHelper
+  def create_collection (title, creator, description)
+    page.should have_content 'Create New Collection'
+    fill_in('Title', with: title)
+    fill_in('Creator', with: creator)
+    fill_in('Description', with: description)
+    click_button("Create Collection")
+    page.should have_content 'Items in this Collection'
+    page.should have_content(title)
+    page.should have_content(creator)
+    page.should have_content(description)
+  end
+end
+
+RSpec.configure do |config|
+  config.include CollectionsHelper
+end
